backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
google-search-results==2.4.2
